Add companies house service
Prior to implementing a validator for companies house number this changes adds the companies house service to the project.

The service will be used to make calls to companies house to validate the number provided e.g. is it recognised and does it belong to an active company.

The code for the service has come from [WCR renewals](https://github.com/DEFRA/waste-carriers-renewals) and [WEX engine](https://github.com/DEFRA/waste-exemptions-engine), where it was essentially duplicated.

This change is the first step in moving the companies house number validation to a shared gem and thus removing the duplication.