Add companies house number validator
This is the first actual validator we are adding to the gem, and it covers validating a company registration number.

Limited companies (LTD) and Limited Liability Partnerships (LLP) have to be registered with [Companies House](https://beta.companieshouse.gov.uk) so as part of our services we like to validate any number submitted.

The first step is ensuring a value has been provided. We then check the format is valid before making a call to Companies House to confirm its recognised and the company is 'active'.

We already added the service object which handles talking to companies house in a previous commit. This change adds the validator itself which makes use of `CompaniesHouseService`.

This change also incorporates a few other things

- adding [I18n](https://github.com/ruby-i18n/i18n) as a dev dependency to ensure translations work when running tests, so we can confirm the right message is being pulled through
- some housekeeping in `spec_helper.rb`, including removing unnecessary `require "spec_helper" calls in the tests (this is handled by `.rspec`)