Fix configuration ready for hosting in engine
The intention of this gem is that it will actually be used in an engine rather than directly in the host app.

After some initial tests we found that the current setup prevented this. These changes resolve the issues, notably

- Add the missing require link between `lib/defra_ruby_validators/engine.rb` and `lib/defra_ruby_validators/validators.rb`
- Fix where the config should sit (`lib/defra_ruby_validators/validators.rb`)
- Make it so you do not have to call `DefraRubyValidators.configure` before calling `DefraRubyValidators.configuration` to avoid an error
- Move VCR setup in rspec to its own support file
- Move DefraRubyValidators setup in rspec to its own support file

These changes mean the gem can be mounted within another engine, and the rspec setup more closely matches initializing the gem in a host.