Fix gemspec file name
With the gem being called **defra_ruby_validators** the gemspec file should also be called this. As we had left the kebab case name following the generation of the gem, it errored when we came to try and deploy it (`Gemspec file not found: defra_ruby_validators.gemspec`).

This fixes the gemspec name to match the gem name.