Add a true/false validator
This adds a new validator to the gem, specifically for checking if a value is either 'true' or 'false.

In the Waste Exemptions service there are a couple of pages where we want the user to provide a simple **yes/no** answer to a question.

Because of this we hold the values as booleans in the database, however this gets translated in the view as setting the radio button values to 'true' and 'false'.

When the answer is submitted is when the validator is expected to be used hence it validates whether the submitted value is either 'true' or 'false'.