Move and modify validators.rb
This is quite a big change. Previously validators.rb was also handling the configuration for the gem but as we're following the exporters template, this will now be handled by a separate class.

The requires will also come back in once we start to move and refactor them to use the new namespace.