Fix missing translations when loaded
The issue was this project was not loading as an engine within the WEX engine (when testing) because engine.rb was not getting called.

First in the WEX engine we updated the require statement to read as `require "defra_ruby/validators". This stopped the engine exploding when loading but did mean none of the translations got loaded. This was tracked down to the fact that statement was pointing to `lib/defra_ruby/validators/validators.rb` and this does not include the code for loading the engine locales.

The fix was to update the require statement to read as `require "defra_ruby/validators/engine"`. However this meant that `lib/defra_ruby.rb` was meaningless. It did nothing anymore.

So to make the API a little cleaner, particularly from the context of the host engine, we renamed it to `defra_ruby_validators.rb` and this means the existing `require "defra_ruby_validators"` in the WEX engines `lib/waste_exemptions_engine/engine.rb` still works.