Refactor namespace to be DefraRuby::Validators (#6)
We recently built the gem [Defra-ruby-exporters](https://github.com/DEFRA/defra-ruby-exporters). It allows us to reuse the code we have built up for creating CSV files and storing them in AWS S3.

As part of that we decided that as we are starting to build up a series of things under the `defra-ruby` namespace, we should use a better namespace pattern in our gems/engines.

Hence in the exporters gem we use `DefraRuby::Exporters`. We want to be consistent across all our gems. So before we add too much more functionality, this change updates the namespace to `DefraRuby::Validators` from `DefraRubyValidators`.

It also includes a few bits of housekeeping

- add the missing `.env.example` file
- update VCR usage to manage it in before and after hooks, rather than using the code block syntax
- fix, tidy and bring travis.yml inline with other projects (specifically code coverage was not being reported)