Make rspec support structure consistent
We now have an agreed structure for rspec in our projects, where all additional tools and dependencies used within the test suite are initialized in separate files in the support folder rather than being referred to directly in the spec_helper.

This gem was put together before that agreement was put in place so this change just brings it inline with our other projects.