Add base validator class (#10)
This is the first step in aligning the existing companies house validator with those in the [Waste Exemptions Engine](https://github.com/DEFRA/waste-exemptions-engine) prior to migrating them all into this gem.

They now make use of a base validator class which handles the error message for all validators. This change brings it in, though it tweaks the way the I18n translation for the error is identified.

In the engine currently they are identified by the object being validated e.g. business_type_form. In the gem it'll be easier to base the keys on the validator being used instead.

To support this we have added a private method that returns the name of the class without its namespace and updated the translation file to match.

One side effect is that we have to use a camelcase key rather than underscore in the locale files. We could have stuck with underscore, but we would then have had to add a further step to the `class_name()` method. This step would have involved copying the implementation of the [Rails ActiveSupport underscore](https://api.rubyonrails.org/classes/ActiveSupport/Inflector.html#method-i-underscore) or bringing in ActiveSupport as a dependency.

This seemed overkill to avoid having a camelcase key (`class_name()` will return `CompaniesHouseNumberValidator`) hence the update to the locale file.