Remove companies house number help text
There was an intention that this gem could also store some of the additional content used in a view such as hints and helper text, as this would enable sharing and avoid duplication.

In reality this is not being used, and if we did go this route we'd look at a more appropriate gem to do it in.

Hence this change removes it from the locale file and the README.