Add use of attribute to error message (#12)
When we copied the base validator from WEX in [PR #10](https://github.com/DEFRA/defra-ruby-validators/pull/10) we dropped the use of the `attribute` param because it didn't appear to be needed.

However with the benefit of hindsight we've realised that there are some validators which are handling more than one attribute, e.g. validating a person's name using first and last name attributes.

Those validators are not part of the gem yet, but they soon will be so it makes sense to correct our refactoring of the `BaseValidator.error_message` to handle attributes.

In the locales we've chosen to drop the key `errors` because with [PR #11](https://github.com/DEFRA/defra-ruby-validators/pull/11) our locales will now only contain error messages. So this updates the Companies House validation messages to work with the use of an attribute, we do a straight swap for `errors:` to `company_no:`.