Refactor Companies House Validator unit tests
This change refactors the existing companies house unit tests to make use of shared examples.

The shared examples have been directly taken from the waste exemptions engine where work was done to remove the duplication in the validation tests.

This is the final step in preparing this project for migrating the rest of the validators in WEX to this gem.