Make attribute optional in error message
We are looking to add our next validation which will be a TrueFalseValidator (referred to as YesNoValidator in other projects though the valid values are true and false hence the rename).

It only has the one attribute to be checked, and that attribute will change depending on what the validator has been assigned to.

So as far as our the translations are concerned, we don't need or want to know what the attribute is called. This means we can't rely on it existing as a key in the locale yml file, and therefore cannot use it in our call to I18n.

So this change updates the base validator to make the params to `error_message()` both named and optional. It can then determine how to format the call to I18n based on whether the attribute param has been passed in.

N.B. This also updates the VCR cassettes.