Add new TrueFalseValidator
This is essentially a copy and paste of the `YesNoValidator` from [waste-exemptions-engine](https://github.com/DEFRA/waste-exemptions-engine).

The intention is to replace it with this one in our validators gem, however when we migrated it across we spotted that the values it actually checks against are `true` and `false`. Therefore it made sense as part of the migration to rename the validator based on what it's actually checking against.

We also bring in our first [concern](https://api.rubyonrails.org/classes/ActiveSupport/Concern.html) as there will be other 'selection' based validators to be migrated, and the concern allows us to share code between them.