Add position validator plus chars & length concern
This is essentially a copy and paste of the validator from [waste-exemptions-engine](https://github.com/DEFRA/waste-exemptions-engine).

Adding it here means we can begin to share the validator across our services.

To support adding it we also copied over the `CanValidateLength` concern.

Finally as part of the refactoring when copying it over we took the opportunity to create the `CanValidateCharacters` concern. On review we could see that the position and person_name validators were duplicating `value_has_no_invalid_characters?()`.