Remove dependency on activesupport (#23)
We brought in ActiveSupport as that is what [waste-exemptions](https://github.com/DEFRA/waste-exemptions-engine) used for its validation concerns.

However it raises a Hakiri warning

- CVE2015-3227
- Specially crafted XML documents can cause applications to raise a `SystemStackError` and potentially cause a denial
- Workarounds: Use an XML parser that is not impacted by this problem, such as Nokogiri or LibXML

Because our other projects are Rails based they automatically have Nokogiri and as such this appears to alleviate the issue with ActiveSupport.

However as just a gem, we don't want to solve the problem by adding an unnecessary dependency. Hence this change which drops ActiveSupport and reformats the 'concerns' to just be vanilla modules.