Add grid reference validator
This is essentially a copy and paste of the validator from [waste-exemptions-engine](https://github.com/DEFRA/waste-exemptions-engine).

Adding it here means we can begin to share the validator across our services.

This also includes adding the [os_map_ref](https://github.com/DEFRA/os-map-ref) gem which allows us to validate that a given grid reference is a valid coordinate.