Ensure we define attributes in concern methods (#31)
For validators that share a concern, we have to define the attribute incase there is an error. If there is the concern needs to know the attrubute so it can then use it to locate the correct error message via the locales.

We have made it clear in the README what attribute to define for each validator, but we want to be sure we are specifying the correct one in validators with a concern.

So this change swaps out `attribute` for `:my_attribute_name` where it is needed.