Enable the gem to use custom validation error messages (#33)
This is to support https://eaflood.atlassian.net/browse/RUBY-493

Each validator in the gem comes with I18n error messages. However, in some cases, we'd like to be able to override these with a custom message. We can do this by accepting the value of options[:message] if it's provided.

This PR also fixes dotenv usage in the specs. This turned out to be an issue with load order. defra_ruby_validators.rb was running before dotenv.rb, then throwing an error because the env vars hadn't been set yet.