Enable gem to receive multiple custom error messages
Previously the gem was set up to receive a custom error message for a validator. However, some validators produce different errors depending on the result - for example, a missing value produces one error while a too-long value produces another. We want to be able to provide custom errors for all circumstances.

This requires some refactoring of the tests as well.