Use attribute correct name during validation
As part of some renaming work going on in the WEX engine forms, we found out that the validator gem was hard-coding attribute names in some cases.
This fix will make sure that the error is added to the record with the correct attribute name passed by the validation code in the forms.