Make error messages attribute agnostic (#42)
https://eaflood.atlassian.net/browse/RUBY-561

We recently modified the validator to not expect specific attribute names and just accept the attribute name that was passed in.

However, our default error messages still relied on a specific attribute name. This was fine for validators which are only used on a single attribute with a specific attribute name, but for validators that were used on lots of attributes (specifically the TrueFalseValidator) it resulted in a translation error.

This commit changes our error message adding to not bother with specific attributes, so we should get the correct default error no matter what.