Change true false validator to match true and false objects
In the current forms we are matching agains strings because we use to pass to the validator what is coming from the view.
Rails, however, would convert the string to a proper boolean value when the value is assigned to the object, and in the current form refactoring we would first assign the value to rails and then validate it. Hence we will be back dealing with proper booleans, and we don't need this validator to deal with strings but with booleans.