Fix bin/console
Sometimes its useful to be able to load the gem and interact with the classes inorder to test and debug the code.

A great way to do this is the `bin/console` file found in most gems. However the version in this gem raises an error when you try and us it.

This change covers resolving the error.