Swap irb for pry
When you load the console the default is to open `irb` with your project loaded. However our standard is to use `pry` in other projects.

In fact switch is so common that the code to use pry is automatically added when you create the gem. Its just commented out.

We bring in the gem `pry-byebug`, which in turn brings in gem. So we have everything we need to switch to using our preferred REPL.