Fix bin/console (#47)
Sometimes it's useful to be able to load the gem and interact with the classes in order to test and debug the code.

A great way to do this is the `bin/console` file found in most gems. However, the version in this gem raises an error when you try and use it.

This change covers resolving the error.

Also when you load the console the default is to open `irb` with your project loaded. However, our standard is to use `pry`.

In fact, switching is so common that the code to use `pry` is automatically added when you create the gem. It's just commented out.

We bring in the gem `pry-byebug`, which in turn brings in gem. So we have everything we need to switch to using our preferred REPL. Hence we make that change here as well.